Adds a `FunctionCalling` processor to handle tool use.
This new processor orchestrates a loop where a generative model is called, any function calls returned by the model are executed, and the results are fed back into the model for subsequent turns, up to a maximum number of calls. It includes an `ExecuteFunctionCall` part processor to run the functions.

PiperOrigin-RevId: 822090787