Rollback making ProcessorPart to inherit from genai.Part
The initial transition has been made in a faith that frameworks can unify on a shared content representation and avoid unnecessary conversions. However it requires that representation to be flexible enough to accomodate needs of its users. Unfortunately this attempt has failed and adding anything not directly needed for Gemini models to genai.Part is nearly impossible.

PiperOrigin-RevId: 828318675