Adds async function calling.
Async function calls are non-blocking function calls that are particularly well suited for realtime agents. They allow long-running operations to be run on the background without interrupting the main realtime agents. When such functions return a stream of parts and are implemented with processors, async function calls provide a simple mechanism to implement multi-agents.

PiperOrigin-RevId: 828839605