SQL Alchemy based persistent cache implementation. Using it with sqlite backend is the easiest option for local workflows.
Why SQLite: performance wise it is a terrible choice and any sane deployment should use something like memcached or redis. And SQL (e.g. postgres) is just an overkill. However sqlite is a zero-conf option that does not require any daemon running. So it is very easy to use for local workflows when we don't care that much about performance.

PiperOrigin-RevId: 829383146