Add aiosqlite to [dev] dependencies.
This package is required to use SqlCache with sqlite and in particular to run SqlCache tests. Not adding it to the main library dependencies as concrete SQL engine to use is a choice by each user. SQL is good for prototyping and local development, but not suitable for server-like deployments.

PiperOrigin-RevId: 829432301