Refactor: Extract live WebSocket server logic used for the live commentator into a reusable module for any live processor.
The WebSocket server implementation for AI Studio connections has been moved from `commentator_ais.py` to a new `server.py` module. This allows other live processors to easily run as WebSocket servers by providing a factory function to `server.run_server`. `commentator_ais.py` is updated to use this new `server` module.

This CL also fixes a bug in the bytes() method of ProcessorPart that was failing whenever the mimetype was not text.

PiperOrigin-RevId: 839233309